requirements.txt
Flask
requests
python-dotenv
flask-login
Flask-SQLAlchemy
psycopg2-binary
